+++ myems-api/reports/combinedequipmentcost.py
@@ -497,8 +497,7 @@ class Reporting:
                                            " WHERE equipment_id = %s "
                                            "     AND energy_category_id = %s "
                                            "     AND start_datetime_utc >= %s "
-                                           "     AND start_datetime_utc < %s "
-                                           " ORDER BY start_datetime_utc ",
+                                           "     AND start_datetime_utc < %s ",
                                            (associated_equipment['id'],
                                             energy_category_id,
                                             reporting_start_datetime_utc,